Change deduplication method
Now the hash is calculated of the beginning N bytes of each file.
Deduplication table stores a vector of vectors of strings for each hash.
The strings are paths, and vectors of strings collect paths to files
that are duplicate with each other.
This is because files can begin similarly and have the same hash,
but the rest of the file can be different.
Additionally, it can be specified if the given directory is searched
for duplicates recursively.