Add const and constexpr, fix base class
Added const type qualifiers and constexpr specifiers to valid places.

Added a virtual destructor and protected special member functions to
the class DirectoryOperation. This prevents slicing caused by
polymorphic copying.

Moved the type alias declaration of DedupTable from header to source
file. It is not needed in the public interface.