Add grouping by file size
Now files are first collected into an unordered_map with their size
as key and path as value. Files with unique size are then removed
from deduplication, because they can't have duplicates.

Deduplication now reads the previous map of files for paths to check.

Dedup_table now has an outer unordered_map that allows indexing by
file size, so only files with the same size are compared with each
other.

Also, a bug with exception handling in utilities.cpp was fixed,
a test was added and the possibility to skip initial file count was
removed.

The code was greatly simplified and performance improved.