Fix which file is kept in non-prompting actions
The order of paths given in the command line used to have no effect
on which file was kept in non-prompting delete or link actions. Only
the last modification time determined it. This was despite the help
text saying otherwise. Now each file knows from which path it was
found. Duplicates are sorted first by that path, and if it is the same,
by their last modification time. The first one is kept.